// Web/ClientApp/services/models/tree-species.ts
export interface TreeGenusModel {
    commonGenus: string;
    genus: string;
}

//export interface TreeSpecyModel {
//    common: string;
//    common_genus: string;
//    genus: string;
//    species: string;
//    telaBotanicaTaxon: string;
//    floreProperties: string[];
//}



//export interface SpecyInfoModel {
//    _id: string;
//    usage: string;
//    images: string[];
//    habitat: string;
//    description: string;
//}