// Web/ClientApp/services/generated/species-model.ts
//----------------------
// <auto-generated>
//     Generated by Ozytis Model Generator  DO NOT EDIT!
// </auto-generated>
//----------------------

export interface SpeciesModel {
    commonGenus: string;
    commonSpeciesName: string;
    genus: string;
    id: string;
    speciesName: string;
    telaBotanicaTaxon: string;
    floraKeyValues: string[];
}